Moving the adapters files to their correct places
