+++ app/models/storage/adapters.rb
@@ -1,0 +1,3 @@
+module Storage
+  module Adapters; end
+end
